==> 1_Epibolus_pulchripes_weight_loss_analysis.Rmd <==
---
title: "Microbiome curing through antibiotics"
subtitle: "Weight loss analysis"
author: "Nweze Julius"
date: "`r Sys.Date()`"
link-citations: yes
csl: fems-microbiology-ecology.csl
output:
  rmarkdown::html_document:
    toc: true
    toc_float: true
    toc_depth: 5
    keep_md: true
    number_sections: false
    highlight: "pygments"
    theme: "flatly"
    dev: "png"
    df_print: "kable"
    fig_caption: true
    code_folding: "show"
editor_options: 
  chunk_output_type: console
---

```{r libraries, include=F}
# Load libraries
#.libPaths(c('~/R/library', .libPaths())) # Uncomment if you have no write access to R path

repo <- "http://cran.wu.ac.at"
lib.loc <- Sys.getenv("R_LIBS_USER")

update.packages(
    lib.loc, 
    repos = repo,
    ask = FALSE
)

.cran_libs <- c(
  "knitr", # A General-Purpose Package for Dynamic Report Generation in R
  "kableExtra", # Construct Complex Table with 'kable' and Pipe Syntax
  "rmarkdown", # Dynamic Documents for R
  "extrafont", # for extra figure fonts
  "tidyverse", # for dplyr forcats ggplot2 readr tibble
  "grid", # The Grid Graphics Package
  "magrittr", # pipes
  "scales", # Generic plot scaling methods
  "svglite", # for svg files
  "vegan",
  "AICcmodavg", # To compare these models and find which one is the best fit for the data
  "car", # Companion to Applied Regression
  "rcompanion", #Functions to Support Extension Education Program Evaluation
  "multcomp", # Simultaneous Inference in General Parametric Models 
  "nlme", # Fit Linear Model Using Generalized Least Squares
  "ggResidpanel", # Panels and Interactive Versions of Diagnostic Plots using 
  "lsmeans", # Least-Squares Means
  "survminer", # For Survival Analysis
  "ranger",
  "readr",
  "report", # automatically produces reports of models and data frames according to best practices guidelines
  "lme4", # Mixed-effects models
  "forcats",
  "dplyr",
  "ggfortify",
  "gtsummary", 
   "cmprsk",  # Competing Risks Packages
  "sm",  # Smooth survival plot - quantile of survival
  "prodlim",
  "pec"
) 

.inst <- .cran_libs %in% installed.packages()
if (any(!.inst)) {
   install.packages(.cran_libs[!.inst],
                    repos = repo,
                    lib = lib.loc)
}

.bioc_libs <- c(
  #"multtest", #Resampling-based multiple hypothesis testing
)

.bioc_inst <- .bioc_libs %in% installed.packages()
if (any(!.bioc_inst)) {
   if (!requireNamespace("BiocManager", quietly = TRUE))
   install.packages("BiocManager")
   BiocManager::install(ask = F, lib = lib.loc)  # upgrade bioC packages
   BiocManager::install(.bioc_libs[!.bioc_inst], ask = F, lib = lib.loc)
}

.local_libs <- c()

.inst <- names(.local_libs) %in% installed.packages()
if (any(!.inst)) {
   install.packages(paste0("~/R/", .local_libs[!.inst]) ,repos = NULL, type = "source", lib = lib.loc)
}

.github_libs <- c()

.github_lib_names <- stringr::str_replace(.github_libs, ".*/(.*)$", "\\1")

.github_inst <- .github_lib_names %in% installed.packages()
if (any(!.github_inst)) {
  devtools::install_github(.github_libs[!.github_inst],
                           lib = lib.loc,
                           dependencies = TRUE)
}

# Load packages into session, and print package version
(loaded.libs <- sapply(c(.cran_libs, .bioc_libs, names(.local_libs), .github_lib_names), require, character.only = TRUE))
if (!all(loaded.libs)) {stop(paste("Package(s):", names(loaded.libs[loaded.libs == FALSE]), "could not be loaded"))}
sapply(c(.cran_libs, .bioc_libs, names(.local_libs), .github_lib_names), packageVersion)
```

```{r style settings, include=F}
options(width = 90, knitr.table.format = "html") 
opts_chunk$set(
  warning = FALSE,
  message = FALSE,
  cache = TRUE,
  dev = "svglite",
  fig.ext = "svg",
  dpi = 300,
#  fig.width = 12,
#  fig.height = 8,
  cache.path = "Weight_loss_cache/",
  fig.path = "Weight_loss_figs/"
)
f_name <- "DejaVu Sans" #sub("\\s//", "", f_name)
f_size <- 14
font_import(pattern = "DejaVuSans", prompt = FALSE)
loadfonts() # registers fonts
theme_set(theme_bw(base_size = f_size, base_family = f_name)) # set theme for plots
pom4 <- ggpomological:::pomological_palette[c(2, 9, 3, 1)] # set colours
```

```{r functions, include=F}
PlotLmResid <- function(lm.df, which = c(1:6), mfrow = c(3, 2)){

  if (length(levels(as.factor(lm.df$.fitted))) < 10) {# if number of unique x values is <10 just draw a line through the means
    smoother <- stat_summary(fun.y = mean, colour = "red", geom = "line")
  } else smoother <- stat_smooth(method = "loess", geom = "smooth", se = FALSE, colour = "firebrick", size = 1)
  
  # residuals vs fitted
  g1 <- ggplot(lm.df, aes(.fitted, .resid)) +
    geom_point()  +
    smoother + 
    geom_hline(yintercept = 0, linetype = 2, size = .2) +
    scale_x_continuous("Fitted Values") +
    scale_y_continuous("Residual") +
    labs(title = "Residuals vs Fitted")
  
  # normal qq
  a <- quantile(lm.df$.stdresid, c(0.25, 0.75), na.rm = TRUE)
  b <- qnorm(c(0.25, 0.75))
  slope <- diff(a)/diff(b)
  int <- a[1] - slope * b[1]
  g2 <- ggplot(lm.df, aes(sample = .stdresid)) +
    stat_qq() +
    geom_abline(slope = slope, intercept = int, colour = "firebrick", size = 1) +
      scale_x_continuous("Theoretical Quantiles") +
      scale_y_continuous("Standardized Quantiles") +
      labs(title = "Normal Q-Q")
 
  # scale-location
  g3 <- ggplot(lm.df, aes(.fitted, sqrt(abs(.stdresid)))) +
    geom_point() +
    smoother +
    scale_x_continuous("Fitted Values") +
    scale_y_continuous("Root of Standardized Residuals") +
    labs(title = "Scale-Location")
 
  # residuals vs leverage
  g4 <- ggplot(lm.df, aes(factors, .stdresid)) +
    geom_point() +
    smoother +
    geom_hline(yintercept = 0, linetype = 2, size = .2) +
    scale_x_continuous("Factor Level Combinations") +
    scale_y_continuous("Standardized Residuals") +
    labs(title = "Residuals vs Factor Levels")
 
#   # cook's distance
#   g5 <-  ggplot(lm.df, aes(rows, .cooksd, ymin=0, ymax=.cooksd)) +
#     geom_point() + geom_linerange() +
#     scale_x_continuous("Observation Number") +
#     scale_y_continuous("Cook's distance") +
#     labs(title="Cook's Distance")  
  
  # cooksd vs leverage
  g5 <- ggplot(lm.df, aes(factors, .cooksd)) +
    geom_point() +
    smoother +
    scale_x_continuous("Factor Level Combinations") +
    scale_y_continuous("Cook's distance") +
    labs(title = "Cook's dist vs Leverage")
  
  # g6 <- PlotACF(lm.df)
  bw <- diff(range(lm.df$.resid)) / (2 * IQR(lm.df$.resid) / length(lm.df$.resid) ^ (1/3))
  sshist <- function(x){ # optimise bins
  # 2006 Author Hideaki Shimazaki
  # Department of Physics, Kyoto University
  # shimazaki at ton.scphys.kyoto-u.ac.jp
	N <- 2 : 100
	C <- numeric(length(N))
	D <- C

	for (i in 1:length(N)) {
		D[i] <- diff(range(x)) / N[i]

		edges = seq(min(x), max(x), length=N[i])
		hp <- hist(x, breaks = edges, plot=FALSE)
		ki <- hp$counts

		k <- mean(ki)
		v <- sum((ki-k) ^ 2) / N[i]

		C[i] <- (2 * k-v) / D[i] ^ 2	#Cost Function
	}

	idx <- which.min(C)
	optD <- D[idx]

	bins <- seq(min(x), max(x), length=N[idx])
	# h = hist(x, breaks = bins)
	# rug(x)

	return(bins)
  }
  
  bins <- sshist(lm.df$.resid)
  g6 <- ggplot(lm.df, aes(.resid)) + 
    geom_histogram(breaks = bins)
 
  plots <- list(g1, g2, g3, g4, g5, g6)
 
  # making the plots
  grid.newpage()
 
  if (prod(mfrow) > 1) {
    mypos <- expand.grid(1:mfrow[1], 1:mfrow[2])
    mypos <- mypos[with(mypos, order(Var1)), ]
    pushViewport(viewport(layout = grid.layout(mfrow[1], mfrow[2])))
    formatter <- function(.){}
  } else {
    mypos <- data.frame(matrix(1, length(which), 2))
    pushViewport(viewport(layout = grid.layout(1, 1)))
    formatter <- function(.) {
      .dontcare <- readline("Hit <Return> to see next plot: ")
      grid.newpage()
    }
  }
 
  j <- 1
  for (i in which) {
    formatter()
    print(plots[[i]], vp = viewport(layout.pos.row = mypos[j, ][1], layout.pos.col = mypos[j, ][2]))
    j <- j + 1
  }
}

PlotACF <- function(lm.df){
  ## generate ACF plot for lm and lme
  # compute acf without plotting
  acz <- acf(lm.df$.resid, plot = F)
  # convert to data frame
  acd <- data.frame(lag = acz$lag, acf = acz$acf)
  # use data frame for ggplot
  ggplot(acd, aes(lag, acf)) + 
    geom_bar(colour = "black", fill = "black", stat = "identity", position = "dodge", width = 0.01) +
    geom_point(colour = "black") +
    geom_hline(yintercept = c(0.05, -0.05), linetype = "dashed") +
    geom_hline(yintercept = 0)
}
```
[roey.angel@bc.cas.cz](mailto: roey.angel@bc.cas.cz)  

## Weight loss analysis
Analysis of animal weight loss over time in response to antibiotics treatment.
Species tested: *E. pulchripes* and *G. connexa*.

### *E. pulchripes* analysis 

```{r load E. pulchripes data, cache = T}
Epibolus_weights <- read_csv("1_Epibolus_weight.csv",
                             trim_ws = TRUE,
                             col_names = TRUE) %>% 
mutate(across(c("Code",
                  "Treatment",
                  "Replicate",
                  "Individual"),
                ~ factor(.))) %>% 
mutate(across(c("Date"), ~ as.Date(., format = "%d/%m/%Y"))) %>% 
mutate(Treatment = fct_relevel(Treatment, "Control", "Sterile", "2X Diluted", "Undiluted"))
```

#### Plot time series

**Plot time series of all replicates**
```{r Epibolus plot all weights, cache=TRUE}
ggplot(Epibolus_weights, aes(x = Day, y = `Weight (g)`, colour = Treatment, group = Code)) + 
  geom_point() +
  geom_line() +
  scale_colour_manual(values = pom4, name = "") 
```

**Plot time series of means**
```{r Epibolus plot mean+SE weights, cache=TRUE}
Epibolus_weights %>%  # the name of the new data frame and the data frame to be summarised
  drop_na() %>% # drop rows with missing values
  group_by(Treatment, Day) %>%   # the grouping variable
  summarise(`Mean weight (g)` = mean(`Weight (g)`),  # calculates the mean of each group
            StD = sd(`Weight (g)`), # calculates the standard deviation of each group
            n = n(),  # calculates the sample size per group
            SE = sd(`Weight (g)`)/sqrt(n())) -> # calculates the standard error of each group
  Epibolus_weights_mean

# Create plot
ggplot(Epibolus_weights_mean, aes(x = Day, y = `Mean weight (g)`)) +
  geom_point(aes(colour = Treatment), size = 6) +
  geom_ribbon(aes(
    ymin = `Mean weight (g)` - SE,
    ymax = `Mean weight (g)` + SE,
    fill = Treatment
  ),
  alpha = 1 / 4) +
  geom_line(aes(colour = Treatment)) +
  labs(y = "Weight (g)") +
  scale_colour_manual(values = pom4, name = "") +
  scale_fill_manual(values = pom4, name = "") 
```

```{r Epibolus table mean+SE weights, cache=TRUE}
Epibolus_weights_mean %>% # output a table of group means
  kable(., digits = c(0, 1, 1, 1, 0, 1)) %>%
  kable_styling(bootstrap_options = c("striped", "hover", "condensed", "responsive"), full_width = F)
```

#### Anova tests
**Compare time 0 weights between groups**
```{r Epibolus t0 weight aov, cache=T}
Epibolus_weights %>% 
  filter(., Day  == 0) %>% # select only day 0 entries 
  aov(`Weight (g)` ~ Treatment, .) -> # fit an anova model
  aov_t0
  
print(Anova(aov_t0)) # print results

aov_t0_df <- fortify(aov_t0) # make a data frame for plotting
factor.combinations <-
  as.numeric(factor(aov_t0_df$Treatment),
             levels = unique(as.character(paste(aov_t0_df$Treatment)))
  ) # needed for "residuals vs leverage
mod_data_df <-
  cbind(aov_t0_df,
        rows = 1:nrow(aov_t0_df),
        factors = factor.combinations)

PlotLmResid(mod_data_df) # plot residuals
```

The ANOVA model indicates no significant different in the weights of the animals at time 0

**Post-hoc tests for time 0**
```{r Epibolus t0 weight post-hoc, cache=T}
TukeyHSD(aov_t0) # Compute Tukey Honest Significant Differences
```



#### Anova tests
**Compare time 7 weights between groups**
```{r Epibolus t7 weight aov, cache=T}
Epibolus_weights %>% 
  filter(., Day  == 7) %>% # select only day 7 entries 
  aov(`Weight (g)` ~ Treatment, .) -> # fit an anova model
  aov_t7
  
print(Anova(aov_t7)) # print results

aov_t7_df <- fortify(aov_t7) # make a data frame for plotting
factor.combinations <-
  as.numeric(factor(aov_t7_df$Treatment),
             levels = unique(as.character(paste(aov_t7_df$Treatment)))
  ) # needed for "residuals vs leverage
mod_data_df <-
  cbind(aov_t7_df,
        rows = 1:nrow(aov_t7_df),
        factors = factor.combinations)

PlotLmResid(mod_data_df) # plot residuals
```

The ANOVA model indicates no significant different in the weights of the animals at time 7

**Post-hoc tests for time 7**
```{r Epibolus t7 weight post-hoc, cache=T}
TukeyHSD(aov_t7) # Compute Tukey Honest Significant Differences
```
Also in the post-hoc tests no pair of treatment was found to have a significant difference in weights.


#### Anova tests
**Compare time 14 weights between groups**
```{r Epibolus t14 weight aov, cache=T}
Epibolus_weights %>% 
  filter(., Day  == 14) %>% # select only day 14 entries 
  aov(`Weight (g)` ~ Treatment, .) -> # fit an anova model
  aov_t14
  
print(Anova(aov_t14)) # print results

aov_t14_df <- fortify(aov_t14) # make a data frame for plotting
factor.combinations <-
  as.numeric(factor(aov_t14_df$Treatment),
             levels = unique(as.character(paste(aov_t14_df$Treatment)))
  ) # needed for "residuals vs leverage
mod_data_df <-
  cbind(aov_t14_df,
        rows = 1:nrow(aov_t14_df),
        factors = factor.combinations)

PlotLmResid(mod_data_df) # plot residuals
```

The ANOVA model indicates no significant different in the weights of the animals at time 14

**Post-hoc tests for time 14**
```{r Epibolus t14 weight post-hoc, cache=T}
TukeyHSD(aov_t14) # Compute Tukey Honest Significant Differences
```
Also in the post-hoc tests no pair of treatment was found to have a significant difference in weights.


#### Anova tests
**Compare time 21 weights between groups**
```{r Epibolus t21 weight aov, cache=T}
Epibolus_weights %>% 
  filter(., Day  == 21) %>% # select only day 21 entries 
  aov(`Weight (g)` ~ Treatment, .) -> # fit an anova model
  aov_t21
  
print(Anova(aov_t21)) # print results

aov_t21_df <- fortify(aov_t21) # make a data frame for plotting
factor.combinations <-
  as.numeric(factor(aov_t21_df$Treatment),
             levels = unique(as.character(paste(aov_t21_df$Treatment)))
  ) # needed for "residuals vs leverage
mod_data_df <-
  cbind(aov_t21_df,
        rows = 1:nrow(aov_t21_df),
        factors = factor.combinations)

PlotLmResid(mod_data_df) # plot residuals
```

The ANOVA model indicates no significant different in the weights of the animals at time 21

**Post-hoc tests for time 21**
```{r Epibolus t21 weight post-hoc, cache=T}
TukeyHSD(aov_t21) # Compute Tukey Honest Significant Differences
```
Also in the post-hoc tests no pair of treatment was found to have a significant difference in weights.


#### Anova tests
**Compare time 28 weights between groups**
```{r Epibolus t28 weight aov, cache=T}
Epibolus_weights %>% 
  filter(., Day  == 28) %>% # select only day 28 entries 
  aov(`Weight (g)` ~ Treatment, .) -> # fit an anova model
  aov_t28
  
print(Anova(aov_t28)) # print results

aov_t28_df <- fortify(aov_t28) # make a data frame for plotting
factor.combinations <-
  as.numeric(factor(aov_t28_df$Treatment),
             levels = unique(as.character(paste(aov_t28_df$Treatment)))
  ) # needed for "residuals vs leverage
mod_data_df <-
  cbind(aov_t28_df,
        rows = 1:nrow(aov_t28_df),
        factors = factor.combinations)

PlotLmResid(mod_data_df) # plot residuals
```

The ANOVA model indicates no significant different in the weights of the animals at time 28


**Post-hoc tests for time 28**
```{r Epibolus t28 weight post-hoc, cache=T}
TukeyHSD(aov_t28) # Compute Tukey Honest Significant Differences
```
Also in the post-hoc tests no pair of treatment was found to have a significant difference in weights.



#### Anova tests
**Compare time 35 weights between groups**
```{r Epibolus t35 weight aov, cache=T}
Epibolus_weights %>% 
  filter(., Day  == 35) %>% # select only day 35 entries 
  aov(`Weight (g)` ~ Treatment, .) -> # fit an anova model
  aov_t35
  
print(Anova(aov_t35)) # print results

aov_t35_df <- fortify(aov_t35) # make a data frame for plotting
factor.combinations <-
  as.numeric(factor(aov_t35_df$Treatment),
             levels = unique(as.character(paste(aov_t35_df$Treatment)))
  ) # needed for "residuals vs leverage
mod_data_df <-
  cbind(aov_t35_df,
        rows = 1:nrow(aov_t35_df),
        factors = factor.combinations)

PlotLmResid(mod_data_df) # plot residuals
```

The ANOVA model indicates no significant different in the weights of the animals at time 35

**Post-hoc tests for time 35**
```{r Epibolus t35 weight post-hoc, cache=T}
TukeyHSD(aov_t35) # Compute Tukey Honest Significant Differences
```
Also in the post-hoc tests no pair of treatment was found to have a significant difference in weights.


#### Anova tests
**Compare time 42 weights between groups**
```{r Epibolus t42 weight aov, cache=T}
Epibolus_weights %>% 
  filter(., Day  == 42) %>% # select only day 42 entries 
  aov(`Weight (g)` ~ Treatment, .) -> # fit an anova model
  aov_t42
  
print(Anova(aov_t42)) # print results

aov_t42_df <- fortify(aov_t42) # make a data frame for plotting
factor.combinations <-
  as.numeric(factor(aov_t42_df$Treatment),
             levels = unique(as.character(paste(aov_t42_df$Treatment)))
  ) # needed for "residuals vs leverage
mod_data_df <-
  cbind(aov_t42_df,
        rows = 1:nrow(aov_t42_df),
        factors = factor.combinations)

PlotLmResid(mod_data_df) # plot residuals
```

The ANOVA model indicates no significant different in the weights of the animals at time 42

**Post-hoc tests for time 42**
```{r Epibolus t42 weight post-hoc, cache=T}
TukeyHSD(aov_t42) # Compute Tukey Honest Significant Differences
```
Also in the post-hoc tests no pair of treatment was found to have a significant difference in weights at dat 42.


*Subtracting the initial weight from the final weight*

**Compare time 35 weights between groups by subtracting the initial weight**
To do that I'll first subtract the weight at time 35 from time 0 for each individual
```{r Epibolus t35 weight aov, cache=T}
Epibolus_weights %>% 
  filter(., Day  == 0 | Day == 35) %>% # select only day 0 and 35 entries 
  pivot_wider(id_cols = c(Code, Treatment, Replicate, Individual), names_from = Day, values_from = `Weight (g)`) %>% # make a wide format table
  mutate(`Weight diff. (g)` = `35` - `0`) %>% # subtract day 0 from day 35 weight
  aov(`Weight diff. (g)` ~ Treatment, .) -> # fit anova
  aov_t35
  
print(Anova(aov_t35)) # print results

aov_t35_df <- fortify(aov_t35) # make a data frame for plotting

factor.combinations <-
  as.numeric(factor(aov_t35_df$Treatment),
             levels = unique(as.character(paste(aov_t35_df$Treatment)))
  ) # needed for "residuals vs leverage
mod_data_df <-
  cbind(aov_t35_df,
        rows = 1:nrow(aov_t35_df),
        factors = factor.combinations)

PlotLmResid(mod_data_df) # plot residuals
```
The analysis indicates a significant difference in weights at day 35

**Post-hoc tests for time 35**
```{r Epibolus t35 weight post-hoc, cache=T}
TukeyHSD(aov_t35)
```

Pairs: Sterile-Control, 2X diluted-Control and Undiluted-Control have significant differences between them but there's no difference between antibiotic treatment and sterile food!




**Compare time 42 weights between groups by subtracting the initial weight**
To do that I'll first subtract the weight at time 42 from time 0 for each individual
```{r Epibolus t42 weight aov, cache=T}
Epibolus_weights %>% 
  filter(., Day  == 0 | Day == 42) %>% # select only day 0 and 42 entries 
  pivot_wider(id_cols = c(Code, Treatment, Replicate, Individual), names_from = Day, values_from = `Weight (g)`) %>% # make a wide format table
  mutate(`Weight diff. (g)` = `42` - `0`) %>% # subtract day 0 from day 42 weight
  aov(`Weight diff. (g)` ~ Treatment, .) -> # fit anova
  aov_t42
  
print(Anova(aov_t42)) # print results

aov_t42_df <- fortify(aov_t42) # make a data frame for plotting

factor.combinations <-
  as.numeric(factor(aov_t42_df$Treatment),
             levels = unique(as.character(paste(aov_t42_df$Treatment)))
  ) # needed for "residuals vs leverage
mod_data_df <-
  cbind(aov_t42_df,
        rows = 1:nrow(aov_t42_df),
        factors = factor.combinations)

PlotLmResid(mod_data_df) # plot residuals
```

The analysis indicates a significant difference in weights at day 42

**Post-hoc tests for time 42**
```{r Epibolus t42 weight post-hoc, cache=T}
TukeyHSD(aov_t42)
```

Pairs: Sterile-Control, 2X diluted-Control and Undiluted-Control have significant differences between them but there's no difference between antibiotic treatment and sterile food!




#### Time-series analysis
So far we've only taken into account the last time point, however we can analyse the entire time series to see if there is a difference in trends.

**Fit a linear mixed-effects model**
```{r Epibolus time-series analysis}
Epibolus_weights %>% 
  drop_na() %>% # drop missing entries
  rename(Weight = `Weight (g)`) -> # rename column 
  Epibolus_weights4model # data frame for modeling

model.a <- 
  lme(Weight ~ Treatment + Day + Treatment * Day, 
      random = ~ Day | Individual,
      data = Epibolus_weights4model) # determine autocorrelations/lag

autocorr.lme <- ACF(model.a,
                form = ~ Day | Individual)[2, 2]

# Fit a linear mixed-effects model 
lme(Weight ~ Treatment + Day + Treatment * Day, 
    data = Epibolus_weights4model,
    random = ~Day|Individual,
    correlation = corAR1(form = ~ Day | Individual, 
                         value = autocorr.lme),
    method = "REML") ->
lme.fit

summary(lme.fit)
   
anova(lme.fit)

report::report(lme.fit)
```
The model indicates some minor effect of the treatment on certain days


**Plot diagnostic plots for the lme model**
```{r Epibolus diag plots lme, cache=T}
resid_panel(lme.fit)

# # We can see the QQ-plot:
print(qqnorm(lme.fit, abline = c(0, 1), pch = 16))

# We saw 5 outliers. Let us identify these observations in out database. First, let us obtain the standardized residuals
res=resid(lme.fit, type="p")

# Then we look for the observations with an absolute standardized residual greater than 2:
which(abs(res) >2)


# We see that there are, in fact, 14 observations that lead to extreme residuals. Actually, these observations come from the individuals labelled 23, 31, 34, 3, 29 and 40. Let us inspect the individual with
# ID = 23:
res[Epibolus_weights4model$Individual == 23]

# 23         23         23         23         23         23         23 
# 2.8382062 -1.8069709 -0.8794832 -0.7661836  0.3060486  0.2987277  0.1647554

# Let us see these observations
subset(Epibolus_weights4model, Individual == 23)

# Code  Weight Treatment Replicate Date         Day Individual Status
#  <fct>  <dbl> <fct>     <fct>     <date>     <dbl> <fct>       <dbl>
# 1 3C      5.96 Undiluted 3         2019-11-05     0 23              0
# 2 3C      5.14 Undiluted 3         1941-11-12     7 23              0
# 3 3C      5.24 Undiluted 3         1941-11-19    14 23              0
# 4 3C      5.21 Undiluted 3         1941-11-26    21 23              0
# 5 3C      5.34 Undiluted 3         2019-12-03    28 23              0
# 6 3C      5.28 Undiluted 3         2019-12-10    35 23              0
# 7 3C      5.21 Undiluted 3         2019-12-17    42 23              0

# The weight was decreasing and increasing on different days.
# The weight was decreasing and increasing on different days. It  appears as an outlier in the residuals. The problem with outliers is that sometimes they have an important influence in the model estimates
# We can check this by adjusting the model again, but excluding these individuals. If the two models give similar estimates, we will conclude that the outliers are simple anomalies in the model. If the two models give similar estimates, we will conclude that the outliers are simple anomalies in the model. However, if they give very different estimates, we would consider invalid the model with outliers.

# Let us adjust the model again but excluding individuals 23, 31, 34, 3, 29 and 40:

Epibolus_weights4model2 = subset(Epibolus_weights4model, Individual!=22 & Individual!=31 & Individual!=34 & Individual!=3 & Individual!=29 & Individual!=40)
mod7 = lme(Weight ~ Day + Treatment+ Day*Treatment, dat = Epibolus_weights4model2, random=~ Day|Individual)
summary(mod7)

# The standard deviation estimate of the random effect is 1.22 and 0.0038, whereas in the model with the complete dataset (the best chosen model) it was 1.17 and 0.0035. Regarding the random error, now we have a standard deviation of 0.12 in this model, whereas in the preivous model (the best chosen model) it was 0.17. Thus, the change in these two variabilities is small. If we look at the fixed effects, we conclude that the estimates barely change from one model to another, and the interaction term is still statistically significant. In conclusion, the six animals affecting the residuals do not have an important impact in the model estimates so we decide to keep the first model.

```

Check if including random effects term is justified
```{r Epibolus rand, cache=T}
# is the random effect meaningful?
Epibolus_weights4model %>% 
  gls(Weight ~ Treatment + Day + Treatment * Day,
      data = .,
      method = "REML") ->
  gls.fixed.fit

anova(lme.fit, gls.fixed.fit)
```
Yes

*Model significance*
Calculate a p-value and pseudo R-squared value for the model by comparing it to a null model with no fixed effects
```{r Epibolus model p-value and post-hoc tests}
Epibolus_weights4model %>% 
  lme(Weight ~ 1,
      random = ~Day|Individual,
      data = .) ->
  model.null

nagelkerke(lme.fit,
           model.null)

marginal <- lsmeans(lme.fit, ~Treatment:Day)

cld(marginal,
    alpha   = 0.05,
    Letters = letters,     ### Use lower-case letters for .group
    adjust  = "tukey") %>%     ###  Tukey-adjusted comparisons
  kable(., digits = c(0, 1, 2, 1, 1, 1, 1, 0)) %>%
  kable_styling(bootstrap_options = c("striped", "hover", "condensed", "responsive"), full_width = F)
```





# Linear mixed model with random effect model
```{r Epibolusrandom effect model}
Epibolus_weights %>% 
  drop_na() %>% # drop missing entries
  rename(Weight = `Weight (g)`) -> # rename column 
  Epibolus_REM # data frame for modeling

# Model estimation
# Slopes and intercepts varying per group
mod1 <- lme(Weight ~ Day*Treatment, random = ~1|Individual, data = Epibolus_REM)
mod2 <- lme(Weight ~ Day + Treatment, random = ~1|Individual, data = Epibolus_REM)
mod3 <- lme(Weight ~ Day + Treatment + Day*Treatment, random = ~1|Individual, data = Epibolus_REM)


# Perform anova function: # The Akaike information criterion (AIC) is a mathematical method for evaluating how well a model fits the data it was generated from. In statistics, AIC is used to compare different possible models and determine which one is the best fit for the data. 
anova(mod1, mod2, mod3)

# mod2 has the lowest AIC and it is he best
summary(mod2)
# The importance of a random effect can be assessed by comparing its variance to that of the random error. In this case, the standard deviation of the random effect b0i is approximately 0.18 times higher than than of the random error, thus this random effect is important


# More formally, we can evaluate the null hypothesis σ0 = 0. If we do not reject it, we will conclude that it is not necessary to include this random effect in the model. First, we should estimate the model with no random effect using the function gls

mod4 = gls(Weight ~ Day + Treatment, data = Epibolus_REM)
summary(mod4)

# Once the model with no random effect is estimated, the hypothesis test should be performed with the function anova:
anova(mod4, mod2)

# The p-value is lower than 0.05 (the usual type I error used in hypothesis tests), so we have enough evidence to say that the variance of the random effect is different from 0.

# Then we think about adding a random effect in the Day slope:
mod5 <- lme(Weight ~ Day + Treatment, random = ~Day|Individual, data = Epibolus_REM)
summary(mod5)

# The standard deviation of the random effect b2i is very small compared to that of the random effect b0i. If we test the null hypothesis of σ2 = 0
anova(mod5, mod2)

# We reject the null hypothesis and so we should include this effect in our model.
# In conclusion, our best model is the one that includes a random effect in the intercept and slope, and then a fixed effect in the slope.

# Now let us evaluate the fixed effects
#  First of all, we should test if there is an interaction between Day and Treatment. Remember that the interaction effect implies the following: the slope between the Weight (our response variable) and Day is different depending on the Treatment. Thus, we need to test the null hypothesis β3 = 0 versus the alternative β3 6 = 0. To solve this contrast we may use the test shown in the summary table of the fixed effects in mod3:


mod6 <- lme(Weight ~ Day + Treatment+ Day*Treatment, random = ~Day|Individual, data = Epibolus_REM)
summary(mod6)

anova(mod6, mod5)


# The p-value corresponding to the interaction term is lower than 0.05, so we reject the null hypothesis. Thus, there is a significant interaction between Day and Treatment.


# The general model that we have obtained can be written as follows:
# Sterile
# Weight = 4.859261 - 0.772571 + 0.003401 * Day - 0.006111 * Day = 4.08669 -0.00271 * Day (D14; 4.05)

# 2X Diluted
# Weight = 4.859261 - 0.487531 + 0.003401 * Day - 0.008187 * Day = 4.37173 - 0.004786 * Day (D14; 4.30)

# Undiluted
# Weight = 4.859261 - 0.727212 + 0.003401 * Day - 0.008672 * Day = 4.132049 - 0.005271 * Day (D14; 4.06)

#mod5# Weight = 5.089914 - 1.071452 - 0.002251 * Day = 4.018462 - 0.002251 * Day (D14; 3.986)

# Control 
# Weight = 4.859261 + 0.003401 * Day (D14; 4.812)

# The slopes for antibiotics-treated groups were less than that for Control or Sterile group. That is, the weight in Days is less. Moreover, the Weight increases, in mean, 0.003401 g per day in the Control group, and decreases 0.00271 in Sterile, 0.004786 in 2X Diluted and 0.005271 in Undiluted.
# The random effect in the intercept can be interpreted as the variability of the initial weight between Control and other treatments. This variability is represented by a standard deviation of 1.17 g.

# Run anova for mod6
anova(mod6)
# automatically produces reports of models and data frames according to best practices guidelines
report::report(mod6)




# Model validation: Once we have chosen a model for our data, the next step is to check if the assumptions of the model are met. We need to validate the following assumptions: independence of the random errors, normality of residuals and presence of outliers.

# We can check the independence of the random errors using a dispersion plot of the standardized residuals versus the predicted values:
print(plot(mod6, pch=16, which=1))

# In this plot we see that the behaviour of the residuals is, in general, right. Most of them take values between -2 and 2 and there are only 15 observations with a standardized residual (in absolute value) greater than 2. Moreover, we do not see any grouping of observations nor trends.

# We can see the QQ-plot:
print(qqnorm(mod6, abline = c(0, 1), pch = 16))


# We saw 5 outliers. Let us identify these observations in out database. First, let us obtain the standardized residuals
res=resid(mod6, type="p")

# Then we look for the observations with an absolute standardized residual greater than 2:
which(abs(res) >2)

# We see that there are, in fact, 15 observations that lead to extreme residuals. Actually, these observations come from the individuals labelled 23, 31, 34, 3, 29 and 40. Let us inspect the individual with
# ID = 23:
res[Epibolus_REM$Individual == 23]

# 23         23         23         23         23         23         23 
# 2.8382062 -1.8069709 -0.8794832 -0.7661836  0.3060486  0.2987277  0.1647554

# Let us see these observations
subset(Epibolus_REM, Individual == 23)

# Code  Weight Treatment Replicate Date         Day Individual Status
#  <fct>  <dbl> <fct>     <fct>     <date>     <dbl> <fct>       <dbl>
# 1 3C      5.96 Undiluted 3         2019-11-05     0 23              0
# 2 3C      5.14 Undiluted 3         1941-11-12     7 23              0
# 3 3C      5.24 Undiluted 3         1941-11-19    14 23              0
# 4 3C      5.21 Undiluted 3         1941-11-26    21 23              0
# 5 3C      5.34 Undiluted 3         2019-12-03    28 23              0
# 6 3C      5.28 Undiluted 3         2019-12-10    35 23              0
# 7 3C      5.21 Undiluted 3         2019-12-17    42 23              0

# The weight was decreasing and increasing on different days.

# If we look at subject 31:
res[Epibolus_REM$Individual == 31]

#        31         31         31         31         31         31         31 
# 2.3152401  2.1562872 -3.4547100 -2.4315826 -0.9500974 -0.4938861  2.9114952

# Let us see these observations
subset(Epibolus_REM, Individual == 31)

#  Code  Weight Treatment  Replicate Date         Day Individual Status
#   <fct>  <dbl> <fct>      <fct>     <date>     <dbl> <fct>       <dbl>
# 1 4A      5.22 2X Diluted 1         2019-11-05     0 31              0
# 2 4A      5.16 2X Diluted 1         1949-11-12     7 31              0
# 3 4A      4.19 2X Diluted 1         1949-11-19    14 31              0
# 4 4A      4.32 2X Diluted 1         1949-11-26    21 31              0
# 5 4A      4.53 2X Diluted 1         2019-12-03    28 31              0
# 6 4A      4.56 2X Diluted 1         2019-12-10    35 31              0
# 7 4A      5.09 2X Diluted 1         2019-12-17    42 31              0


# The weight was decreasing and increasing on different days. It  appears as an outlier in the residuals. The problem with outliers is that sometimes they have an important influence in the model estimates
# We can check this by adjusting the model again, but excluding these individuals. If the two models give similar estimates, we will conclude that the outliers are simple anomalies in the model. If the two models give similar estimates, we will conclude that the outliers are simple anomalies in the model. However, if they give very different estimates, we would consider invalid the model with outliers.

# Let us adjust the model again but excluding individuals 23, 31, 34, 3, 29 and 40:

Epibolus_REM2 = subset(Epibolus_REM, Individual!=22 & Individual!=31 & Individual!=34 & Individual!=3 & Individual!=29 & Individual!=40)
mod7 = lme(Weight ~ Day + Treatment+ Day*Treatment, dat = Epibolus_REM2, random=~ Day|Individual)
summary(mod7)

# The standard deviation estimate of the random effect is in mod7 1.22, whereas in the model with the complete dataset (mod6) it was 1.17. Regarding the random error, now we have a standard deviation of 0.12 in mod7, whereas in the preivous model (mod6) it was 0.17. Thus, the change in these two variabilities is small. If we look at the fixed effects, we conclude that the estimates barely change from one model to another, and the interaction term is still statistically significant. In conclusion, the six animals affecting the residuals do not have an important impact in the model estimates so we decide to keep the first model.


```
