==> 7_Epibolus_pulchripes_faecal_bacterial_colony_counts.Rmd <==
---
title: "Microbiome curing through antibiotics"
subtitle: "Colony counts analysis"
author: "Nweze Julius"
date: "`r Sys.Date()`"
link-citations: yes
csl: fems-microbiology-ecology.csl
output:
  rmarkdown::html_document:
    toc: true
    toc_float: true
    toc_depth: 5
    keep_md: true
    number_sections: false
    highlight: "pygments"
    theme: "flatly"
    dev: "png"
    df_print: "kable"
    fig_caption: true
    code_folding: "show"
editor_options: 
  chunk_output_type: console
---

```{r libraries, include=F}
# Load libraries
#.libPaths(c('~/R/library', .libPaths())) # Uncomment if you have no write access to R path

repo <- "http://cran.wu.ac.at"
lib.loc <- Sys.getenv("R_LIBS_USER")

update.packages(
    lib.loc, 
    repos = repo,
    ask = FALSE
)

.cran_libs <- c(
  "knitr", # A General-Purpose Package for Dynamic Report Generation in R
  "kableExtra", # Construct Complex Table with 'kable' and Pipe Syntax
  "rmarkdown", # Dynamic Documents for R
  "extrafont", # for extra figure fonts
  "tidyverse", # for dplyr forcats ggplot2 readr tibble
  "grid", # The Grid Graphics Package
  "magrittr", # pipes
  "scales", # Generic plot scaling methods
  "svglite", # for svg files
  "vegan",
  "psych",
  "car", # Companion to Applied Regression
  "rcompanion", #Functions to Support Extension Education Program Evaluation
  "multcomp", # Simultaneous Inference in General Parametric Models 
  "nlme", # Fit Linear Model Using Generalized Least Squares
  "ggResidpanel", # Panels and Interactive Versions of Diagnostic Plots using 
  "lsmeans" # Least-Squares Means
) 

.inst <- .cran_libs %in% installed.packages()
if (any(!.inst)) {
   install.packages(.cran_libs[!.inst],
                    repos = repo,
                    lib = lib.loc)
}

.bioc_libs <- c(
  #"multtest", #Resampling-based multiple hypothesis testing
)

.bioc_inst <- .bioc_libs %in% installed.packages()
if (any(!.bioc_inst)) {
   if (!requireNamespace("BiocManager", quietly = TRUE))
   install.packages("BiocManager")
   BiocManager::install(ask = F, lib = lib.loc)  # upgrade bioC packages
   BiocManager::install(.bioc_libs[!.bioc_inst], ask = F, lib = lib.loc)
}

.local_libs <- c()

.inst <- names(.local_libs) %in% installed.packages()
if (any(!.inst)) {
   install.packages(paste0("~/R/", .local_libs[!.inst]) ,repos = NULL, type = "source", lib = lib.loc)
}

.github_libs <- c()

.github_lib_names <- stringr::str_replace(.github_libs, ".*/(.*)$", "\\1")

.github_inst <- .github_lib_names %in% installed.packages()
if (any(!.github_inst)) {
  devtools::install_github(.github_libs[!.github_inst],
                           lib = lib.loc,
                           dependencies = TRUE)
}

# Load packages into session, and print package version
(loaded.libs <- sapply(c(.cran_libs, .bioc_libs, names(.local_libs), .github_lib_names), require, character.only = TRUE))
if (!all(loaded.libs)) {stop(paste("Package(s):", names(loaded.libs[loaded.libs == FALSE]), "could not be loaded"))}
sapply(c(.cran_libs, .bioc_libs, names(.local_libs), .github_lib_names), packageVersion)
```

```{r style settings, include=F}
options(width = 90, knitr.table.format = "html") 
opts_chunk$set(
  warning = FALSE,
  message = FALSE,
  cache = TRUE,
  dev = "svglite",
  fig.ext = "svg",
  dpi = 300,
#  fig.width = 12,
#  fig.height = 8,
  cache.path = "Colony_counts_cache/",
  fig.path = "Colony_counts_figs/"
)
f_name <- "DejaVu Sans" #sub("\\s//", "", f_name)
f_size <- 14
font_import(pattern = "DejaVuSans", prompt = FALSE)
loadfonts() # registers fonts
theme_set(theme_bw(base_size = f_size, base_family = f_name)) # set theme for plots
pom4 <- ggpomological:::pomological_palette[c(2, 9, 3, 1)] # set colours
```

```{r functions, include=F}
PlotLmResid <- function(lm.df, which = c(1:6), mfrow = c(3, 2)){

  if (length(levels(as.factor(lm.df$.fitted))) < 10) {# if number of unique x values is <10 just draw a line through the means
    smoother <- stat_summary(fun.y = mean, colour = "red", geom = "line")
  } else smoother <- stat_smooth(method = "loess", geom = "smooth", se = FALSE, colour = "firebrick", size = 1)
  
  # residuals vs fitted
  g1 <- ggplot(lm.df, aes(.fitted, .resid)) +
    geom_point()  +
    smoother + 
    geom_hline(yintercept = 0, linetype = 2, size = .2) +
    scale_x_continuous("Fitted Values") +
    scale_y_continuous("Residual") +
    labs(title = "Residuals vs Fitted")
  
  # normal qq
  a <- quantile(lm.df$.stdresid, c(0.25, 0.75), na.rm = TRUE)
  b <- qnorm(c(0.25, 0.75))
  slope <- diff(a)/diff(b)
  int <- a[1] - slope * b[1]
  g2 <- ggplot(lm.df, aes(sample = .stdresid)) +
    stat_qq() +
    geom_abline(slope = slope, intercept = int, colour = "firebrick", size = 1) +
      scale_x_continuous("Theoretical Quantiles") +
      scale_y_continuous("Standardized Quantiles") +
      labs(title = "Normal Q-Q")
 
  # scale-location
  g3 <- ggplot(lm.df, aes(.fitted, sqrt(abs(.stdresid)))) +
    geom_point() +
    smoother +
    scale_x_continuous("Fitted Values") +
    scale_y_continuous("Root of Standardized Residuals") +
    labs(title = "Scale-Location")
 
  # residuals vs leverage
  g4 <- ggplot(lm.df, aes(factors, .stdresid)) +
    geom_point() +
    smoother +
    geom_hline(yintercept = 0, linetype = 2, size = .2) +
    scale_x_continuous("Factor Level Combinations") +
    scale_y_continuous("Standardized Residuals") +
    labs(title = "Residuals vs Factor Levels")
 
#   # cook's distance
#   g5 <-  ggplot(lm.df, aes(rows, .cooksd, ymin=0, ymax=.cooksd)) +
#     geom_point() + geom_linerange() +
#     scale_x_continuous("Observation Number") +
#     scale_y_continuous("Cook's distance") +
#     labs(title="Cook's Distance")  
  
  # cooksd vs leverage
  g5 <- ggplot(lm.df, aes(factors, .cooksd)) +
    geom_point() +
    smoother +
    scale_x_continuous("Factor Level Combinations") +
    scale_y_continuous("Cook's distance") +
    labs(title = "Cook's dist vs Leverage")
  
  # g6 <- PlotACF(lm.df)
  bw <- diff(range(lm.df$.resid)) / (2 * IQR(lm.df$.resid) / length(lm.df$.resid) ^ (1/3))
  sshist <- function(x){ # optimise bins
  # 2006 Author Hideaki Shimazaki
  # Department of Physics, Kyoto University
  # shimazaki at ton.scphys.kyoto-u.ac.jp
	N <- 2 : 100
	C <- numeric(length(N))
	D <- C

	for (i in 1:length(N)) {
		D[i] <- diff(range(x)) / N[i]

		edges = seq(min(x), max(x), length=N[i])
		hp <- hist(x, breaks = edges, plot=FALSE)
		ki <- hp$counts

		k <- mean(ki)
		v <- sum((ki-k) ^ 2) / N[i]

		C[i] <- (2 * k-v) / D[i] ^ 2	#Cost Function
	}

	idx <- which.min(C)
	optD <- D[idx]

	bins <- seq(min(x), max(x), length=N[idx])
	# h = hist(x, breaks = bins)
	# rug(x)

	return(bins)
  }
  
  bins <- sshist(lm.df$.resid)
  g6 <- ggplot(lm.df, aes(.resid)) + 
    geom_histogram(breaks = bins)
 
  plots <- list(g1, g2, g3, g4, g5, g6)
 
  # making the plots
  grid.newpage()
 
  if (prod(mfrow) > 1) {
    mypos <- expand.grid(1:mfrow[1], 1:mfrow[2])
    mypos <- mypos[with(mypos, order(Var1)), ]
    pushViewport(viewport(layout = grid.layout(mfrow[1], mfrow[2])))
    formatter <- function(.){}
  } else {
    mypos <- data.frame(matrix(1, length(which), 2))
    pushViewport(viewport(layout = grid.layout(1, 1)))
    formatter <- function(.) {
      .dontcare <- readline("Hit <Return> to see next plot: ")
      grid.newpage()
    }
  }
 
  j <- 1
  for (i in which) {
    formatter()
    print(plots[[i]], vp = viewport(layout.pos.row = mypos[j, ][1], layout.pos.col = mypos[j, ][2]))
    j <- j + 1
  }
}

PlotACF <- function(lm.df){
  ## generate ACF plot for lm and lme
  # compute acf without plotting
  acz <- acf(lm.df$.resid, plot = F)
  # convert to data frame
  acd <- data.frame(lag = acz$lag, acf = acz$acf)
  # use data frame for ggplot
  ggplot(acd, aes(lag, acf)) + 
    geom_bar(colour = "black", fill = "black", stat = "identity", position = "dodge", width = 0.01) +
    geom_point(colour = "black") +
    geom_hline(yintercept = c(0.05, -0.05), linetype = "dashed") +
    geom_hline(yintercept = 0)
}
```
[roey.angel@bc.cas.cz](mailto: roey.angel@bc.cas.cz)  

## Colony counts analysis
Analysis of colony counts analysis over time in response to antibiotics treatment.
Species tested: *E. pulchripes*

### *E. pulchripes* analysis 

**Load data **
```{r load E. pulchripes data, cache = T}
read_csv("7_Epibolus_colony_counts.csv",
         trim_ws = TRUE,
         col_names = TRUE) %>% # read the csv file
  mutate_at(c("Code",
              "Treatment",
              "Replicate",
              "Individual"),
            ~ (factor(.))) %>% # convert to factor
  mutate_at(c("Date"), ~ as.Date(., "%d/%m/%Y")) %>% # convert to date 
  mutate(Treatment = fct_relevel(Treatment, "Control", "Sterile", "2X Diluted", "Undiluted")) -> # arrange factor levels 4 plotting
  Epibolus_colony_counts

```

#### Plot time series

**Plot time series of all replicates**
```{r plot all colony counts, cache=TRUE}

scientific_10 <- function(x) {
  parse(text=gsub("e", " %*% 10^", scales::scientific_format()(x)))
}
ggplot(Epibolus_colony_counts, aes(x = Day, y = `No. of colonies per pellet`, colour = Treatment, group = Code)) + 
  geom_point() +
  geom_line() +
  scale_colour_manual(values = pom4, name = "") +
  scale_y_continuous(label=scientific_10)
```


**Plot time series of geometric means**
```{r plot geo mean+SE colony counts, cache=TRUE}
Epibolus_colony_counts %>%  # the name of the new data frame and the data frame to be summarised
  filter(`No. of colonies per pellet` > 0) %>%
  drop_na(`No. of colonies per pellet`) %>% # drop rows with missing values
  group_by(Treatment, Day) %>%   # the grouping variable
  summarise(`Mean no. of colonies per pellet` = geometric.mean(`No. of colonies per pellet`),  # calculates the mean of each group
            StD = sd(`No. of colonies per pellet`), # calculates the standard deviation of each group
            n = n(),  # calculates the sample size per group
            SE = sd(`No. of colonies per pellet`)/sqrt(n())) -> # calculates the standard error of each group
  Epibolus_colony_counts_geo_mean


ggplot(Epibolus_colony_counts_geo_mean, aes(x = Day, y = `Mean no. of colonies per pellet`)) +
  geom_point(aes(colour = Treatment), size = 3) +
  geom_ribbon(aes(
    ymin = `Mean no. of colonies per pellet` - SE,
    ymax = `Mean no. of colonies per pellet` + SE,
    fill = Treatment
  ),
  alpha = 1 / 4) +
  geom_line(aes(colour = Treatment)) +
  labs(y = "No. of colonies per pellet") +
  scale_colour_manual(values = pom4, name = "") +
  scale_fill_manual(values = pom4, name = "") +
  scale_y_continuous(label=scientific_10)
```

```{r table mean+SE colony counts, cache=TRUE}
Epibolus_colony_counts_geo_mean %>% # output a table of group means
  kable(., digits = c(0, 1, 1, 1, 0, 1)) %>%
  kable_styling(bootstrap_options = c("striped", "hover", "condensed", "responsive"), full_width = F)
```



#### Anova tests
**Compare time 0 counts between groups**
```{r t0 colony counts aov, cache=T}
Epibolus_colony_counts %>% 
  filter(., Day  == 2) %>% # select only day 0 entries 
  aov(`No. of colonies per pellet` ~ Treatment, .) -> # fit an anova model
  aov_t2
  
print(Anova(aov_t2)) # print results

aov_t2_df <- fortify(aov_t2) # make a data frame for plotting
factor.combinations <-
  as.numeric(factor(aov_t2_df$Treatment),
             levels = unique(as.character(paste(aov_t2_df$Treatment)))
  ) # needed for "residuals vs leverage
mod_data_df <-
  cbind(aov_t2_df,
        rows = 1:nrow(aov_t2_df),
        factors = factor.combinations)

PlotLmResid(mod_data_df) # plot residuals
```

The ANOVA model indicates no significant different in the initial colony counts of the animals at time 2

**Post-hoc tests for time 2**
```{r t2 colony counts post-hoc, cache=T}
TukeyHSD(aov_t2) # Compute Tukey Honest Significant Differences
```

Also in the post-hoc tests no pair of treatment was found to have a significant difference in bacterial load.


**Compare time 35 colony counts between groups**
Day 35 is used because all the animals produced faecal pellets and the no. of colonies were counted.
```{r t35 colony counts aov, cache=T}
Epibolus_colony_counts %>% 
  filter(., Day  == 35) %>% # select only day 35 entries 
  aov(`No. of colonies per pellet` ~ Treatment, .) -> # fit an anova model
  aov_t35
  
print(Anova(aov_t35)) # print results

aov_t35_df <- fortify(aov_t35) # make a data frame for plotting
factor.combinations <-
  as.numeric(factor(aov_t35_df$Treatment),
             levels = unique(as.character(paste(aov_t35_df$Treatment)))
  ) # needed for "residuals vs leverage
mod_data_df <-
  cbind(aov_t35_df,
        rows = 1:nrow(aov_t35_df),
        factors = factor.combinations)

PlotLmResid(mod_data_df) # plot residuals
```
The analysis indicates a significant difference in weights at day 35

**Post-hoc tests for time 35**
```{r t35 weight post-hoc, cache=T}
TukeyHSD(aov_t35)
```
Pairs: Sterile-Control, 2X diluted-Control and Undiluted-Control have significant differences between them but there's no difference between antibiotic treatment and sterlie food!



# Total colony counts
**Plot time series of geometric means for total colony counts**
```{r plot geo mean+SE colony counts, cache=TRUE}
Epibolus_colony_counts %>%
  filter(`Total colonies` > 0) %>%
  drop_na(`Total colonies`) %>%
  group_by(Treatment, Day) %>%
  summarise(
    `Mean total colony counts` = geometric.mean(`Total colonies`),
    StD = sd(`Total colonies`),
    n = n(),
    SE = StD / sqrt(n)
  ) -> Epibolus_total_colony_counts_geo_mean  # Removed arrow assignment for better readability

ggplot(Epibolus_total_colony_counts_geo_mean, aes(x = Day, y = `Mean total colony counts`)) +
  geom_ribbon(
    aes(ymin = `Mean total colony counts` - SE, ymax = `Mean total colony counts` + SE, fill = Treatment),
    alpha = 1 / 4
  ) +
  geom_point(aes(colour = Treatment), size = 3) +
  geom_line(aes(colour = Treatment)) +
  labs(y = "Total colonies") +
  scale_colour_manual(values = pom4, name = "") +
  scale_fill_manual(values = pom4, name = "") +
  scale_y_continuous(label = scientific_format(scale = 10))  # Use `scientific_format` from scales package

```

```{r table mean+SE for total colony counts, cache=TRUE}
Epibolus_total_colony_counts_geo_mean %>% # output a table of group means
  kable(., digits = c(0, 1, 1, 1, 0, 1)) %>%
  kable_styling(bootstrap_options = c("striped", "hover", "condensed", "responsive"), full_width = F)
```


**Fit a linear mixed-effects model**
```{r Epibolus mixed-effects model analysis}
Epibolus_colony_counts %>% 
  drop_na() %>% # drop missing entries
  rename(Tcolonies = `Total colonies`) %>%
  mutate(Tcolonies_log = log10(Tcolonies)) -> # rename column 
  Epibolus_Tcolonies4model # data frame for modeling

# Testing differenly models
model1 <- lme(Tcolonies_log ~ Treatment + Day + Treatment * Day, random = ~ 1 | Individual, data = Epibolus_Tcolonies4model)
model2 <- lme(Tcolonies_log ~ Treatment + Day, random = ~ 1 | Individual, data = Epibolus_Tcolonies4model)
model3 <- lme(Tcolonies_log ~ Treatment * Day, random = ~ 1 | Individual, data = Epibolus_Tcolonies4model)
model4 <- lme(Tcolonies_log ~ Treatment + Day + Treatment * Day, random = ~ Day | Individual, data = Epibolus_Tcolonies4model)

# Compare based on AIC
anova(model1, model2, model3, model4) # model2 has the lowest

summary(model2)


model5 <- gls(Tcolonies_log ~ Treatment + Day + Treatment * Day, data = Epibolus_Tcolonies4model)
model6 <- gls(Tcolonies_log ~ Treatment + Day, data = Epibolus_Tcolonies4model)

anova(model5, model2)
anova(model6, model2)



anova(model2)

report::report(model2)
```



**Plot diagnostic plots for the lme model**
```{r Epibolus diag plots lme, cache=T}
resid_panel(model2)

# # We can see the QQ-plot:
print(qqnorm(model2, abline = c(0, 1), pch = 16))

# We saw 5 outliers. Let us identify these observations in out database. First, let us obtain the standardized residuals
res=resid(model2, type="p")

# Then we look for the observations with an absolute standardized residual greater than 2:
which(abs(res) >2)


# We see that there are, in fact, 14 observations that lead to extreme residuals. Actually, these observations come from the individuals labelled 23, 31, 34, 3, 29 and 40. Let us inspect the individual with
# ID = 23:
res[Epibolus_Tcolonies4model$Individual == 23]

# 23         23         23         23         23         23         23 
# 2.8382062 -1.8069709 -0.8794832 -0.7661836  0.3060486  0.2987277  0.1647554

# Let us see these observations
subset(Epibolus_Tcolonies4model, Individual == 23)

# Code  Weight Treatment Replicate Date         Day Individual Status
#  <fct>  <dbl> <fct>     <fct>     <date>     <dbl> <fct>       <dbl>
# 1 3C      5.96 Undiluted 3         2019-11-05     0 23              0
# 2 3C      5.14 Undiluted 3         1941-11-12     7 23              0
# 3 3C      5.24 Undiluted 3         1941-11-19    14 23              0
# 4 3C      5.21 Undiluted 3         1941-11-26    21 23              0
# 5 3C      5.34 Undiluted 3         2019-12-03    28 23              0
# 6 3C      5.28 Undiluted 3         2019-12-10    35 23              0
# 7 3C      5.21 Undiluted 3         2019-12-17    42 23              0

# The weight was decreasing and increasing on different days.
# The weight was decreasing and increasing on different days. It  appears as an outlier in the residuals. The problem with outliers is that sometimes they have an important influence in the model estimates
# We can check this by adjusting the model again, but excluding these individuals. If the two models give similar estimates, we will conclude that the outliers are simple anomalies in the model. If the two models give similar estimates, we will conclude that the outliers are simple anomalies in the model. However, if they give very different estimates, we would consider invalid the model with outliers.

# Let us adjust the model again but excluding individuals 23, 31, 34, 3, 29 and 40:

Epibolus_weights4model2 = subset(Epibolus_weights4model, Individual!=22 & Individual!=31 & Individual!=34 & Individual!=3 & Individual!=29 & Individual!=40)
mod7 = lme(Weight ~ Day + Treatment+ Day*Treatment, dat = Epibolus_weights4model2, random=~ Day|Individual)
summary(mod7)

# The standard deviation estimate of the random effect is 1.22 and 0.0038, whereas in the model with the complete dataset (the best chosen model) it was 1.17 and 0.0035. Regarding the random error, now we have a standard deviation of 0.12 in this model, whereas in the preivous model (the best chosen model) it was 0.17. Thus, the change in these two variabilities is small. If we look at the fixed effects, we conclude that the estimates barely change from one model to another, and the interaction term is still statistically significant. In conclusion, the six animals affecting the residuals do not have an important impact in the model estimates so we decide to keep the first model.

```







#### Anova tests for total colony counts
**Compare time 2 counts between groups**
```{r t2 colony counts aov, cache=T}
Epibolus_colony_counts %>% 
  filter(., Day  == 2) %>% # select only day 2 entries 
  aov((`Total colonies`) ~ Treatment, .) -> # fit an anova model
  aov_t2
  
print(Anova(aov_t2)) # print results

aov_t2_df <- fortify(aov_t2) # make a data frame for plotting
factor.combinations <-
  as.numeric(factor(aov_t2_df$Treatment),
             levels = unique(as.character(paste(aov_t2_df$Treatment)))
  ) # needed for "residuals vs leverage
mod_data_df <-
  cbind(aov_t2_df,
        rows = 1:nrow(aov_t2_df),
        factors = factor.combinations)

PlotLmResid(mod_data_df) # plot residuals
```

The ANOVA model indicates no significant different in the initial colony counts of the animals at time 2

**Post-hoc tests for time 2**
```{r t0 colony counts post-hoc, cache=T}
TukeyHSD(aov_t2) # Compute Tukey Honest Significant Differences
```

Also in the post-hoc tests no pair of treatment was found to have a significant difference in bacterial load.

```{r t2 colony counts aov, cache=T}
# Building a regression model
Epi_model1_d2 <- Epibolus_colony_counts %>% 
  filter(., Day  == 2) %>% lm(`Total colonies` ~ Treatment, data = .)

print(Anova(Epi_model1_d2, type = "III"))

# checking model assumptions
check_model(Epi_model1_d2)

Epi_emm_mod1_d2 <- emmeans(Epi_model1_d2, as.formula(paste("", "Treatment", sep = " ~ ")), weights = "equal")
      
# Contrasts, pairwise comparisons, tests, and confidence intervals.
Epi_emm_mod_cons1_d2 <- contrast(Epi_emm_mod1_d2, method = "pairwise", adjust = "BH")
print(Epi_emm_mod_cons1_d2)

# The mode shows no statistical significance but before moving on, we need to test the model's assumptions. Obviously the dataset does not fit the assumptions of the model so we cannot use the results and we need to transform the model so that the residuals are normally distributed. One of the easiest ways is to log-transform the data before the analysis.

Epibolus_colony_counts %<>%
  mutate(log_copies = log2(`Total colonies`))

# check the model
Epi_model2_d2 <- Epibolus_colony_counts %>% 
  filter(., Day  == 2) %>% lm(log_copies ~ Treatment, data = .)
print(Anova(Epi_model2_d2, type = "III"))


# checking model assumptions
check_model(Epi_model2_d2)

Epi_emm_mod2_d2 <- emmeans(Epi_model2_d2, as.formula(paste("", "Treatment", sep = " ~ ")), weights = "equal")
      
# Contrasts, pairwise comparisons, tests, and confidence intervals.
Epi_emm_mod_cons2_d2 <- contrast(Epi_emm_mod2_d2, method = "pairwise", adjust = "BH")
print(Epi_emm_mod_cons2_d2)
```

**Compare time 9 total colony counts between groups**
Day 9 is used because all the animals produced faecal pellets and the no. of colonies were counted.
```{r t9 total colony counts aov, cache=T}
Epibolus_colony_counts %>%
  drop_na(`Total colonies`) %>%
  filter(Day == 9) %>%  # Corrected to 'Day' from '.'
  aov(log(`Total colonies`) ~ Treatment,.) -> # Added 'data =' argument
  aov_t9  

print(Anova(aov_t9))  # Use 'Anova' from the 'car' package


aov_t9_df <- fortify(aov_t9) # make a data frame for plotting
factor.combinations <-
  as.numeric(factor(aov_t9_df$Treatment),
             levels = unique(as.character(paste(aov_t9_df$Treatment)))
  ) # needed for "residuals vs leverage
mod_data_df <-
  cbind(aov_t9_df,
        rows = 1:nrow(aov_t9_df),
        factors = factor.combinations)

PlotLmResid(mod_data_df) # plot residuals
```
The analysis indicates a significant difference in weights at day 9

**Post-hoc tests for time 9**
```{r t9 weight post-hoc, cache=T}
TukeyHSD(aov_t9)
```
Pairs: Sterile-Control, 2X diluted-Control and Undiluted-Control have significant differences between them but there's no difference between antibiotic treatment and sterlie food!

```{r t9 colony counts aov, cache=T}
# Building a regression model
Epi_model1_d9 <- Epibolus_colony_counts %>% 
  filter(., Day  == 9) %>% lm(`Total colonies` ~ Treatment, data = .)

print(Anova(Epi_model1_d9, type = "III"))

# checking model assumptions
check_model(Epi_model1_d9)

Epi_emm_mod1_d9 <- emmeans(Epi_model1_d9, as.formula(paste("", "Treatment", sep = " ~ ")), weights = "equal")
      
# Contrasts, pairwise comparisons, tests, and confidence intervals.
Epi_emm_mod_cons1_d9 <- contrast(Epi_emm_mod1_d9, method = "pairwise", adjust = "BH")
print(Epi_emm_mod_cons1_d9)

# The mode shows no statistical significance but before moving on, we need to test the model's assumptions. Obviously the dataset does not fit the assumptions of the model so we cannot use the results and we need to transform the model so that the residuals are normally distributed. One of the easiest ways is to log-transform the data before the analysis.

Epibolus_colony_counts %<>%
  mutate(log_copies = log2(`Total colonies`))

# check the model
Epi_model2_d9 <- Epibolus_colony_counts %>% 
  filter(., Day  == 9) %>% lm(log_copies ~ Treatment, data = .)
print(Anova(Epi_model2_d9, type = "III"))


# checking model assumptions
check_model(Epi_model2_d9)

Epi_emm_mod2_d9 <- emmeans(Epi_model2_d9, as.formula(paste("", "Treatment", sep = " ~ ")), weights = "equal")
      
# Contrasts, pairwise comparisons, tests, and confidence intervals.
Epi_emm_mod_cons2_d9 <- contrast(Epi_emm_mod2_d9, method = "pairwise", adjust = "BH")
print(Epi_emm_mod_cons2_d9)
```


**Compare time 16 total colony counts between groups**
Day 16 is used because all the animals produced faecal pellets and the no. of colonies were counted.
```{r t16 total colony counts aov, cache=T}
Epibolus_colony_counts %>%
  drop_na(`Total colonies`) %>%
  filter(Day == 16) %>%  # Corrected to 'Day' from '.'
  aov(log(`Total colonies`) ~ Treatment,.) -> # Added 'data =' argument
  aov_t16  

print(Anova(aov_t16))  # Use 'Anova' from the 'car' package


aov_t16_df <- fortify(aov_t16) # make a data frame for plotting
factor.combinations <-
  as.numeric(factor(aov_t16_df$Treatment),
             levels = unique(as.character(paste(aov_t16_df$Treatment)))
  ) # needed for "residuals vs leverage
mod_data_df <-
  cbind(aov_t16_df,
        rows = 1:nrow(aov_t16_df),
        factors = factor.combinations)

PlotLmResid(mod_data_df) # plot residuals
```
The analysis indicates a significant difference in weights at day 16

**Post-hoc tests for time 16**
```{r t16 weight post-hoc, cache=T}
TukeyHSD(aov_t16)
```
Pairs: Sterile-Control, 2X diluted-Control and Undiluted-Control have significant differences between them but there's no difference between antibiotic treatment and sterlie food!
```{r t16 colony counts aov, cache=T}
# Building a regression model
Epi_model1_d16 <- Epibolus_colony_counts %>% 
  filter(., Day  == 16) %>% lm(`Total colonies` ~ Treatment, data = .)

print(Anova(Epi_model1_d16, type = "III"))

# checking model assumptions
check_model(Epi_model1_d16)

Epi_emm_mod1_d16 <- emmeans(Epi_model1_d16, as.formula(paste("", "Treatment", sep = " ~ ")), weights = "equal")
      
# Contrasts, pairwise comparisons, tests, and confidence intervals.
Epi_emm_mod_cons1_d16 <- contrast(Epi_emm_mod1_d16, method = "pairwise", adjust = "BH")
print(Epi_emm_mod_cons1_d16)

# The mode shows no statistical significance but before moving on, we need to test the model's assumptions. Obviously the dataset does not fit the assumptions of the model so we cannot use the results and we need to transform the model so that the residuals are normally distributed. One of the easiest ways is to log-transform the data before the analysis.

Epibolus_colony_counts %<>%
  mutate(log_copies = log2(`Total colonies`))

# check the model
Epi_model2_d16 <- Epibolus_colony_counts %>% 
  filter(., Day  == 16) %>% lm(log_copies ~ Treatment, data = .)
print(Anova(Epi_model2_d16, type = "III"))


# checking model assumptions
check_model(Epi_model2_d16)

Epi_emm_mod2_d16 <- emmeans(Epi_model2_d16, as.formula(paste("", "Treatment", sep = " ~ ")), weights = "equal")
      
# Contrasts, pairwise comparisons, tests, and confidence intervals.
Epi_emm_mod_cons2_d16 <- contrast(Epi_emm_mod2_d16, method = "pairwise", adjust = "BH")
print(Epi_emm_mod_cons2_d16)
```


**Compare time 23 total colony counts between groups**
Day 23 is used because all the animals produced faecal pellets and the no. of colonies were counted.
```{r t23 total colony counts aov, cache=T}
Epibolus_colony_counts %>%
  drop_na(`Total colonies`) %>%
  filter(Day == 23) %>%  # Corrected to 'Day' from '.'
  aov(log(`Total colonies`) ~ Treatment,.) -> # Added 'data =' argument
  aov_t23  

print(Anova(aov_t23))  # Use 'Anova' from the 'car' package


aov_t23_df <- fortify(aov_t23) # make a data frame for plotting
factor.combinations <-
  as.numeric(factor(aov_t23_df$Treatment),
             levels = unique(as.character(paste(aov_t23_df$Treatment)))
  ) # needed for "residuals vs leverage
mod_data_df <-
  cbind(aov_t23_df,
        rows = 1:nrow(aov_t23_df),
        factors = factor.combinations)

PlotLmResid(mod_data_df) # plot residuals
```
The analysis indicates a significant difference in weights at day 23

**Post-hoc tests for time 23**
```{r t23 weight post-hoc, cache=T}
TukeyHSD(aov_t23)
```
Pairs: Sterile-Control, 2X diluted-Control and Undiluted-Control have significant differences between them but there's no difference between antibiotic treatment and sterlie food!

```{r t23 colony counts aov, cache=T}
# Building a regression model
Epi_model1_d23 <- Epibolus_colony_counts %>% 
  filter(., Day  == 23) %>% lm(`Total colonies` ~ Treatment, data = .)

print(Anova(Epi_model1_d23, type = "III"))

# checking model assumptions
check_model(Epi_model1_d23)

Epi_emm_mod1_d23 <- emmeans(Epi_model1_d23, as.formula(paste("", "Treatment", sep = " ~ ")), weights = "equal")
      
# Contrasts, pairwise comparisons, tests, and confidence intervals.
Epi_emm_mod_cons1_d23 <- contrast(Epi_emm_mod1_d23, method = "pairwise", adjust = "BH")
print(Epi_emm_mod_cons1_d23)

# The mode shows no statistical significance but before moving on, we need to test the model's assumptions. Obviously the dataset does not fit the assumptions of the model so we cannot use the results and we need to transform the model so that the residuals are normally distributed. One of the easiest ways is to log-transform the data before the analysis.

Epibolus_colony_counts %<>%
  mutate(log_copies = log2(`Total colonies`))

# check the model
Epi_model2_d23 <- Epibolus_colony_counts %>% 
  filter(., Day  == 23) %>% lm(log_copies ~ Treatment, data = .)
print(Anova(Epi_model2_d23, type = "III"))


# checking model assumptions
check_model(Epi_model2_d23)

Epi_emm_mod2_d23 <- emmeans(Epi_model2_d23, as.formula(paste("", "Treatment", sep = " ~ ")), weights = "equal")
      
# Contrasts, pairwise comparisons, tests, and confidence intervals.
Epi_emm_mod_cons2_d23 <- contrast(Epi_emm_mod2_d23, method = "pairwise", adjust = "BH")
print(Epi_emm_mod_cons2_d23)
```



**Compare time 30 total colony counts between groups**
Day 30 is used because all the animals produced faecal pellets and the no. of colonies were counted.
```{r t30 total colony counts aov, cache=T}
Epibolus_colony_counts %>%
  drop_na(`Total colonies`) %>%
  filter(Day == 30) %>%  # Corrected to 'Day' from '.'
  aov(log(`Total colonies`) ~ Treatment,.) -> # Added 'data =' argument
  aov_t30  

print(Anova(aov_t30))  # Use 'Anova' from the 'car' package


aov_t30_df <- fortify(aov_t30) # make a data frame for plotting
factor.combinations <-
  as.numeric(factor(aov_t30_df$Treatment),
             levels = unique(as.character(paste(aov_t30_df$Treatment)))
  ) # needed for "residuals vs leverage
mod_data_df <-
  cbind(aov_t30_df,
        rows = 1:nrow(aov_t30_df),
        factors = factor.combinations)

PlotLmResid(mod_data_df) # plot residuals
```
The analysis indicates a significant difference in weights at day 30

**Post-hoc tests for time 30**
```{r t30 weight post-hoc, cache=T}
TukeyHSD(aov_t30)
```
Pairs: Sterile-Control, 2X diluted-Control and Undiluted-Control have significant differences between them but there's no difference between antibiotic treatment and sterlie food, except 2X Diluted-Sterile

```{r t30 colony counts aov, cache=T}
# Building a regression model
Epi_model1_d30 <- Epibolus_colony_counts %>% 
  filter(., Day  == 30) %>% lm(`Total colonies` ~ Treatment, data = .)

print(Anova(Epi_model1_d30, type = "III"))

# checking model assumptions
check_model(Epi_model1_d30)

Epi_emm_mod1_d30 <- emmeans(Epi_model1_d30, as.formula(paste("", "Treatment", sep = " ~ ")), weights = "equal")
      
# Contrasts, pairwise comparisons, tests, and confidence intervals.
Epi_emm_mod_cons1_d30 <- contrast(Epi_emm_mod1_d30, method = "pairwise", adjust = "BH")
print(Epi_emm_mod_cons1_d30)

# The mode shows no statistical significance but before moving on, we need to test the model's assumptions. Obviously the dataset does not fit the assumptions of the model so we cannot use the results and we need to transform the model so that the residuals are normally distributed. One of the easiest ways is to log-transform the data before the analysis.

Epibolus_colony_counts %<>%
  mutate(log_copies = log2(`Total colonies`))

# check the model
Epi_model2_d30 <- Epibolus_colony_counts %>% 
  filter(., Day  == 30) %>% lm(log_copies ~ Treatment, data = .)
print(Anova(Epi_model2_d30, type = "III"))


# checking model assumptions
check_model(Epi_model2_d30)

Epi_emm_mod2_d30 <- emmeans(Epi_model2_d30, as.formula(paste("", "Treatment", sep = " ~ ")), weights = "equal")
      
# Contrasts, pairwise comparisons, tests, and confidence intervals.
Epi_emm_mod_cons2_d30 <- contrast(Epi_emm_mod2_d30, method = "pairwise", adjust = "BH")
print(Epi_emm_mod_cons2_d30)
```




**Compare time 35 total colony counts between groups**
Day 35 is used because all the animals produced faecal pellets and the no. of colonies were counted.
```{r t35 total colony counts aov, cache=T}
Epibolus_colony_counts %>%
  drop_na(`Total colonies`) %>%
  filter(Day == 35) %>%  # Corrected to 'Day' from '.'
  aov(log(`Total colonies`) ~ Treatment,.) -> # Added 'data =' argument
  aov_t35  

print(Anova(aov_t35))  # Use 'Anova' from the 'car' package


aov_t35_df <- fortify(aov_t35) # make a data frame for plotting
factor.combinations <-
  as.numeric(factor(aov_t35_df$Treatment),
             levels = unique(as.character(paste(aov_t35_df$Treatment)))
  ) # needed for "residuals vs leverage
mod_data_df <-
  cbind(aov_t35_df,
        rows = 1:nrow(aov_t35_df),
        factors = factor.combinations)

PlotLmResid(mod_data_df) # plot residuals
```
The analysis indicates a significant difference in weights at day 35

**Post-hoc tests for time 35**
```{r t35 weight post-hoc, cache=T}
TukeyHSD(aov_t35)
```
Pairs: Sterile-Control, 2X diluted-Control and Undiluted-Control have significant differences between them but there's no difference between antibiotic treatment and sterlie food!

```{r t35 colony counts aov, cache=T}
# Building a regression model
Epi_model1_d35 <- Epibolus_colony_counts %>% 
  filter(., Day  == 35) %>% lm(`Total colonies` ~ Treatment, data = .)

print(Anova(Epi_model1_d35, type = "III"))

# checking model assumptions
check_model(Epi_model1_d35)

Epi_emm_mod1_d35 <- emmeans(Epi_model1_d35, as.formula(paste("", "Treatment", sep = " ~ ")), weights = "equal")
      
# Contrasts, pairwise comparisons, tests, and confidence intervals.
Epi_emm_mod_cons1_d35 <- contrast(Epi_emm_mod1_d35, method = "pairwise", adjust = "BH")
print(Epi_emm_mod_cons1_d35)

# The mode shows no statistical significance but before moving on, we need to test the model's assumptions. Obviously the dataset does not fit the assumptions of the model so we cannot use the results and we need to transform the model so that the residuals are normally distributed. One of the easiest ways is to log-transform the data before the analysis.

Epibolus_colony_counts %<>%
  mutate(log_copies = log2(`Total colonies`))

# check the model
Epi_model2_d35 <- Epibolus_colony_counts %>% 
  filter(., Day  == 35) %>% lm(log_copies ~ Treatment, data = .)
print(Anova(Epi_model2_d35, type = "III"))


# checking model assumptions
check_model(Epi_model2_d35)

Epi_emm_mod2_d35 <- emmeans(Epi_model2_d35, as.formula(paste("", "Treatment", sep = " ~ ")), weights = "equal")
      
# Contrasts, pairwise comparisons, tests, and confidence intervals.
Epi_emm_mod_cons2_d35 <- contrast(Epi_emm_mod2_d35, method = "pairwise", adjust = "BH")
print(Epi_emm_mod_cons2_d35)
```
