==> 21_Epibolus_pulchirpes_Hindgut_Alpha_Diversity.Rmd <==
---
title: "Amplicon Sequence Analyses for Epibolus Hindgut Antibiotic Experiment (using Silva tax)"
subtitle: "Epibolus Hindgut"
author: "Julius Nweze"
date: "`r Sys.Date()`"
output:
  rmarkdown::html_document:
    toc: true
    toc_float: true
    toc_depth: 5
    keep_md: true
    number_sections: false
    highlight: "pygments"
    theme: "flatly"
    dev: "png"
    df_print: "kable"
    fig_caption: true
    code_folding: "show"
editor_options: 
  chunk_output_type: console
---

```{r libraries, include=F}
# Load libraries
#.libPaths(c('~/R/library', .libPaths())) # Uncomment if you have no write access to R path

repo <- "http://cran.wu.ac.at"
lib.loc <- Sys.getenv("R_LIBS_USER")

update.packages(
    lib.loc, 
    repos = repo,
    ask = FALSE
)

.cran_libs <- c(
  "phyloseq", # a set of classes and tools to facilitate the import, storage, analysis, and graphical display of microbiome census data
  "rmarkdown", # Dynamic Documents for R
  "knitr",  # A General-Purpose Package for Dynamic Report Generation in R
  "extrafont", # for extra figure fonts
  "phyloseq.extended", # Various customs functions written to enhance the base functions of phyloseq
  "dplyr",  # filter and reformat data frames
  "gridExtra", # To arrange multiple grid-based plots on a page, and draw tables.
  "knitr", # A General-Purpose Package for Dynamic Report Generation in R
  "kableExtra", # Construct Complex Table with 'kable' and Pipe Syntax
  "rmarkdown", # Dynamic Documents for R
  "extrafont", # for extra figure fonts
  "tidyverse", # for dplyr forcats ggplot2 readr tibble
  "grid", # The Grid Graphics Package
  "magrittr", # pipes
  "scales", # Generic plot scaling methods
  "svglite", # for svg files
  "vegan",  # Ordination methods, diversity analysis and other functions for community and vegetation ecologists
  "car", # Companion to Applied Regression
  "rcompanion", #Functions to Support Extension Education Program Evaluation
  "multcomp", # Simultaneous Inference in General Parametric Models 
  "nlme", # Fit Linear Model Using Generalized Least Squares
  "ggResidpanel", # Panels and Interactive Versions of Diagnostic Plots using 
  "lsmeans", # Least-Squares Means
  "ggplot2",      # graphics
  "tidyr", # necessary to import the data from Excel file
  "phyloseqGraphTest", # Provides functions for graph-based multiple-sample testing and visualization of microbiome data
  "shiny", # makes it easy to build interactive web apps straight from R
  "miniUI", # Designed to work especially well for creating Shiny Gadgets
  "caret", # Classification And REgression Training is a set of functions that attempt to streamline the process for creating predictive models
  "pls", # Partial Least Squares and Principal Component Regression
  "e1071", # Functions for latent class analysis, short time Fourier transform, fuzzy clustering, support vector machines
  "randomForest",
  "ggrepel",
  "dunn.test",
  "parallel",
  "reshape2",
  "FSA",
  "devtools", 
  "PMA", 
#  "structSSI",  
  "ade4",
  "igraph", 
  "rcompanion",
  "ggnetwork",
  "MicrobiotaProcess",
  "intergraph", 
  "scales",
  "microbiome",
  "ggpubr",
  "RColorBrewer",
  "microbiomeutilities",
  "viridis",
  "tibble",
  "cowplot",
  "userfriendlyscience",
 "agricolae", # Statistical Procedures for Agricultural Research
 # "doParallel", # parallel backend for the foreach/%dopar% function
 "BiodiversityR", # Package for Community Ecology and Suitability Analysis
 # "hexbin", # Hexagonal Binning Routines
 # "ggtern", # An Extension to 'ggplot2', for the Creation of Ternary Diagrams
#  "MuMIn" # Multi-Model Inference
  #"stringr",
 # "mctoolsr"
"AnnotationDbi", 
"DESeq2", 
"GO.db", 
"impute", 
"preprocessCore",
"adespatial"
) 

.inst <- .cran_libs %in% installed.packages()
if (any(!.inst)) {
   install.packages(.cran_libs[!.inst],
                    repos = repo,
                    lib = lib.loc)
}

.bioc_libs <- c(
  #"multtest", #Resampling-based multiple hypothesis testing
)

.bioc_inst <- .bioc_libs %in% installed.packages()
if (any(!.bioc_inst)) {
   if (!requireNamespace("BiocManager", quietly = TRUE))
   install.packages("BiocManager")
   BiocManager::install(ask = F, lib = lib.loc)  # upgrade bioC packages
   BiocManager::install(.bioc_libs[!.bioc_inst], ask = F, lib = lib.loc)
}

.local_libs <- c()

.inst <- names(.local_libs) %in% installed.packages()
if (any(!.inst)) {
   install.packages(paste0("~/R/", .local_libs[!.inst]) ,repos = NULL, type = "source", lib = lib.loc)
}

.github_libs <- c()

.github_lib_names <- stringr::str_replace(.github_libs, ".*/(.*)$", "\\1")

.github_inst <- .github_lib_names %in% installed.packages()
if (any(!.github_inst)) {
  devtools::install_github(.github_libs[!.github_inst],
                           lib = lib.loc,
                           dependencies = TRUE)
}


# Load packages into session, and print package version
(loaded.libs <- sapply(c(.cran_libs, .bioc_libs, names(.local_libs), .github_lib_names), require, character.only = TRUE))
if (!all(loaded.libs)) {stop(paste("Package(s):", names(loaded.libs[loaded.libs == FALSE]), "could not be loaded"))}
sapply(c(.cran_libs, .bioc_libs, names(.local_libs), .github_lib_names), packageVersion)
set.seed(123456789)
bootstraps <- 1000
min_lib_size <- 1000
```


```{r style settings, include=F}
graphic_device <- "svglite"
options(width = 90, knitr.table.format = "html") 
opts_chunk$set(
  warning = FALSE,
  message = FALSE,
  cache = TRUE,
  dev = graphic_device,
  fig.ext = "svg",
  #  fig.width=12,
  #  fig.height=8,
  cache.path = "Antibiotics_Sequence_cache/",
  fig.path = "Antibiotics_Sequence_figures/"
)

f_name <- "DejaVu Sans" #sub("\\s//", "", f_name)
f_size <- 12
font_import(pattern = "DejaVuSans", prompt = FALSE)
loadfonts() # registers fonts
theme_set(theme_bw(base_size = f_size, base_family = f_name))
pom4 <- ggpomological:::pomological_palette[c(2, 1, 9, 3)]
pom2 <- c(ggpomological:::pomological_base[[7]], ggpomological:::pomological_palette[[1]])
```




## **Read the data and create Epibolus hindgut phyloseq objects**
```{r load amplicon seq data, cache = T}
# load your data 
tsv_table <- read.table("16_DADA2.seqtab_nochim_decontam.tsv", sep = ",", header = T, quote = "", comment.char = "", stringsAsFactors = F, check.names = FALSE)
tax_table <- read.table("16_DADA2.taxa_silva_decontam.tsv", sep = ",", header = T, quote = "", comment.char = "", stringsAsFactors = F, check.names = FALSE)
sample_table <- read.csv("16_Metadata.csv",  sep = ",", header = T, check.names = FALSE)
```


##*Phyloseq objects need to have row.names*
```{r adding row.names to data, cache = T}

# Define the row names from the ASV column
row.names(tsv_table) <- tsv_table$ASV
 tsv_table <- tsv_table %>% dplyr::select(-c(ASV, MC1_S96, BC1_S103, NC1_S110, MC2_S123, BC2_S130, NC2_S137)) 


row.names(tax_table) <- tax_table$ASV
tax_table <- tax_table %>% dplyr::select(-ASV)
  
row.names(sample_table) <- sample_table$Sample
sample_table <- sample_table[!(sample_table$Sample=="MC1_S96" | sample_table$Sample=="BC1_S103" | sample_table$Sample=="NC1_S110" | sample_table$Sample=="MC2_S123" | sample_table$Sample=="BC2_S130" | sample_table$Sample=="NC2_S137"),]
sample_table <- sample_table %>% dplyr::select(-Sample)
```


*Transform into matrixes tsv and tax tables (sample table can be left as data frame)*
```{r transformation of data to matrices, cache = T} 

tsv_table <- as.matrix(tsv_table)
tax_table <- as.matrix(tax_table)
```


*Transform to phyloseq objects*
```{r pyloseq object, cache = T}


OTU <- otu_table(tsv_table, taxa_are_rows = TRUE)
TAX = phyloseq::tax_table(tax_table)
sample = sample_data(sample_table)

carbom <- merge_phyloseq(sample, OTU, TAX)
carbom

# Visualize data

sample_names(carbom)

rank_names(carbom)

sample_variables(carbom)


summarize_phyloseq(carbom)
```



## *Pre-processing Confirmations*

*Group according to sample sources (Epibolus hingdut and Epibolus faecal pellets)*
```{r group according to sample sources, cache = T}
# Epibolus hingdut
Epi_hindgut <- subset_samples(carbom, sample_names(carbom) == "Hg-E1C_S90" | sample_names(carbom) == "Hg-E1G_S91" | sample_names(carbom) == "Hg-E1H_S92"| sample_names(carbom) == "Hg-E2A_S93" | sample_names(carbom) == "Hg-E2D_S94" | sample_names(carbom) == "Hg-E2H_S95" | sample_names(carbom) == "Hg-E3D_S97" | sample_names(carbom) == "Hg-E3E_S98" | sample_names(carbom) =="Hg-E3J_S99"  | sample_names(carbom) =="Hg-E4B1_S100" | sample_names(carbom) == "Hg-E4B2_S101" | sample_names(carbom) == "Hg-E4C_S102")
Epi_hindgut
```



## **Alpha diversity for Epibolus hindgut dataset**

*Diversity indices in Epibolus hindgut dataset*

*Rarefy without replacement and plot richness*
```{r Epibolus hindgut Rarefy without replacement, cache = T}


# Plot richness for all samples
theme_set(theme_bw())
Epi_hindgut_plot_richness <- plot_richness(Epi_hindgut, x = 'Sample.name', measures = c('Chao1', 'Shannon', 'observed', "Simpson", "InvSimpson", "ACE", "Fisher"), color = 'Treatment') +
  ggtitle("a)") +
  xlab(" ") +
  theme(axis.text.x = element_text(
    angle = 45,
    vjust = 0.9,
    hjust = 0.9
  )) + geom_boxplot()
Epi_hindgut_plot_richness$data$Treatment = factor(Epi_hindgut_plot_richness$data$Treatment, level = c("Control", "Sterile feed", "Antibiotics (2X)", "Antibiotics (1X)"))
Epi_hindgut_plot_richness


# Plot richness for all treatments
theme_set(theme_bw())
Epi_hindgut_plot_richness1 <- plot_richness(Epi_hindgut, x = 'Treatment', measures = c('Chao1', 'Shannon', 'observed', "Simpson", "InvSimpson", "ACE", "Fisher"), color = 'Treatment') +
  ggtitle("a)") +
  xlab(" ") +
  theme(axis.text.x = element_text(
    angle = 90,
    vjust = 0.9,
    hjust = 0.9
  )) + geom_boxplot() +
theme(text = element_text(size = 16),  
axis.text = element_text(size = 10))
Epi_hindgut_plot_richness1$data$Treatment = factor(Epi_hindgut_plot_richness1$data$Treatment, level = c("Control", "Sterile feed", "Antibiotics (2X)", "Antibiotics (1X)"))
Epi_hindgut_plot_richness1

ggsave("Epi_hindgut_richness_its.png", device = "png", path = "Richness_plot")

# Alpha diversity is used to measure the diversity within a sample and answers the question "how many?". It allows you to look at number of different taxa within each sample separately. If a sample has high alpha diversity it contains many organisms. Alpha diversity is how many different taxa are detected in each sample. Alpha diversity looks at each sample and asks how many. There are several methods that can be used to look at and understand alpha diversity. When talking about alpha diversity, we are looking at two things:
# 1. Species richness - a count of the number of different species present in a sample. It does not take into account the abundance of the species or their relative distributions.
# 2. Species evenness - a measure of relative abundance of different species that make up the richness

# Measures for evenness and richness:
# 1. Observed - only considers observed ASVs.
# 2. CHAO1 index - considers not only observed otus, but also non-observed rare species.
# 3. The Simpson diversity index is used to calculate a measure of diversity taking into account the number of taxa as well as the abundance. The simpson index gives more weight to common or dominant species which means a few rare species with only a few representatives will not affect the diversity of the sample.
# 4. The Shannon index summarizes the diversity in the population while assuming all species are represented in a sample and they are randomly sampled. The Shannon index increases as both the richness and evenness of the community increase. 



```


Measure Richness, Shannon, etc, combine with metadta, and then write into a csv file
```{r Epibolus hindgut measure Richness, Shannon, for the phyloseq object, cache = T}
Epi_hindgut_alpha.its <- estimate_richness(Epi_hindgut, measures=c("Observed", "Chao1", "Shannon", "Simpson", "InvSimpson", "ACE", "Fisher"))
Epi_hindgut_alpha.its


# Add SampleID column to the alpha dataframe
Epi_hindgut_alpha.its$Sample <- sample_names(Epi_hindgut)

# Merge alpha and Epibolus hindgut dataset into a single dataframe
Epi_hindgut_metad_alpha_its <- merge(psmelt(Epi_hindgut), Epi_hindgut_alpha.its)

# Write it into a csv file
write.csv(Epi_hindgut_metad_alpha_its, file.path("Epi_hindgut_alpha_its.csv")) # writes out a .csv file that you can open in excel
```
Let's go ahead and write only metrics with mean and sd in table
```{r Epibolus hindgut summarize our models, cache = T}
# Then make a table of means and sd of Observed, ACE, Shannon, Simpson and Fisher and Pielou metrics. This is for publication purposes. You can freely change the columns of diversity metrics to retain. 


# Mean for treatment
Epi_hindgut_alphmean.its = aggregate(Epi_hindgut_metad_alpha_its[,c("Observed", "Chao1", "se.chao1", "Shannon", "Simpson", "InvSimpson", "ACE", "se.ACE", "Fisher")],list(Epi_hindgut_metad_alpha_its$Sample.name, Epi_hindgut_metad_alpha_its$Treatment), mean)

# SD for treatment
Epi_hindgut_alphsd.its = aggregate(Epi_hindgut_metad_alpha_its[, c("Observed", "Chao1", "se.chao1", "Shannon", "Simpson", "InvSimpson", "ACE", "se.ACE", "Fisher")], list(Epi_hindgut_metad_alpha_its$Sample.name, Epi_hindgut_metad_alpha_its$Treatment), sd)

# SE for treatment
se <- function(x) sd(x) / sqrt(length(x))
Epi_hindgut_alphse.its = aggregate(Epi_hindgut_metad_alpha_its[, c("Observed", "Chao1", "se.chao1", "Shannon", "Simpson", "InvSimpson", "ACE", "se.ACE", "Fisher")], list(Epi_hindgut_metad_alpha_its$Sample.name, Epi_hindgut_metad_alpha_its$Treatment), se)

# Rename the columns
colnames(Epi_hindgut_alphmean.its) <- c("Sample.name", "Treatment",  "Observed", "Chao1", "se.chao1", "Shannon", "Simpson", "InvSimpson", "ACE", "se.ACE", "Fisher")
colnames(Epi_hindgut_alphsd.its) <- c("Sample.name", "Treatment", "Observed.sd", "Chao1.sd", "se.chao1.sd", "Shannon.sd", "Simpson.sd", "InvSimpson.sd", "ACE.sd", "se.ACE.sd", "Fisher.sd")
colnames(Epi_hindgut_alphse.its) <- c("Sample.name", "Treatment", "Observed.se", "Chao1.se", "se.chao1.se", "Shannon.se", "Simpson.se", "InvSimpson.se", "ACE.se", "se.ACE.se", "Fisher.se")

    
# Merge and write into a csv file
Epi_hindgut.mean.sd.se <- Reduce(merge, list(Epi_hindgut_alphmean.its, Epi_hindgut_alphsd.its, Epi_hindgut_alphse.its))
options(digits=2)

write.csv(Epi_hindgut.mean.sd.se, file.path("Epi_hindgut_alpha_mean_sd_se.csv"))
```

**Intra-sample variability: Non-parametic test for alpha diversity**
```{r Epibolus hindgut summarize our models, cache = T}
####################### Non-parametic test for alpha diversity
detach(package:plyr)
# Melt the data for plotting
data_melted <- reshape2::melt(Epi_hindgut.mean.sd.se, id.vars = c("Sample.name", "Treatment")) %>%
               subset(variable %in% c("Observed", "Chao1", "Shannon", "Simpson", "InvSimpson", "ACE"))

# Display the melted data
data_melted %>%  kable(digits = c(0, 1, 1, 1, 0, 1)) %>%
  kable_styling(bootstrap_options = c("striped", "hover", "condensed", "responsive"), full_width = F) 


# Function to compute Kruskal-Wallis test and extract p-value
compute_kruskal_pvalue <- function(data) {
  kruskal_result <- kruskal.test(value ~ Sample.name, data = data)
  p_value <- broom::tidy(kruskal_result)$p.value
  return(p_value)
}

# Compute Kruskal-Wallis p-values for each Treatment and variable combination
p_values_list <- data_melted %>%
  group_by(Treatment, variable) %>%
  do(p_value = compute_kruskal_pvalue(.)) %>%
  ungroup()

# Expand the list of p_values
p_values_list_expanded <- p_values_list %>%
  mutate(p_value = unlist(p_value))

# Print the result
print(p_values_list_expanded)

p_values_list_expanded %>%
  mutate(Treatment = factor(Treatment, levels = c("Control", "Sterile feed", "Antibiotics (2X)", "Antibiotics (1X)"))) %>%
  kable(digits = c(0, 1, 1, 1, 0, 1)) %>%
  kable_styling(bootstrap_options = c("striped", "hover", "condensed", "responsive"), full_width = FALSE)

# Convert p-values to alphabetical representation
p_values_list_expanded$alpha_p_value <- factor(ifelse(p_values_list_expanded$p_value > 0.05, "a", "b"))

# Create violin plots with p-value annotations

p_Epi_hindgut <- ggplot(data_melted, aes(x = Treatment, y = value, fill = Treatment)) +
  geom_violin() +
  geom_jitter(width = 0.2, alpha = 0.5) +
  labs(x = "Treatment", y = "Value") +
  #theme_minimal() +
  facet_wrap(~ variable, scales = "free_y") +
  scale_fill_manual(values = pom4) +
  theme(legend.position = "right", text = element_text(size = 26), strip.text.x = element_text(size = 24)) +
  theme(legend.text = element_text(colour = "black", size = 24), axis.text.x = element_text(size = 24, hjust = 1, vjust = 0.5, colour = "black", angle = 90), axis.title = element_text(size = 24), legend.title = element_text(size = 24)) +
  geom_text(data = p_values_list_expanded, aes(x = Treatment, y = Inf, label = alpha_p_value), hjust = 0.25, vjust = 0.4, size = 7)
p_Epi_hindgut

```


**Inter-sample variability: Non-parametic test for alpha diversity**
```{r Epibolus hindgut summarize our models, cache = T}
# Convert 'value' column to numeric
detach(package:plyr)
# Reorder levels of 'Treatment' factor
data_melted$Treatment <- factor(data_melted$Treatment, levels = c("Control", "Sterile feed", "Antibiotics (2X)", "Antibiotics (1X)"))

p_Epi_hindgut <- ggplot(data_melted, aes(x = Treatment, y = value, fill = Treatment)) +
  geom_violin() +
  geom_jitter(width = 0.2, alpha = 0.5) +
  labs(x = "Treatment", y = "Value") +
  #theme_minimal() +
  facet_grid(variable ~ ., scales = "free_y") +
  scale_fill_manual(values = pom4) +
  theme(legend.position = "right", text = element_text(size = 26), strip.text.x = element_text(size = 24)) +
  theme(legend.text = element_text(colour = "black", size = 24), axis.text.x = element_text(size = 24, hjust = 1, vjust = 0.5, colour = "black", angle = 90), axis.title = element_text(size = 24), legend.title = element_text(size = 24)) 
  
p_Epi_hindgut


# Convert 'value' column to numeric
data_melted$value <- as.numeric(data_melted$value)

# Reorder levels of 'Treatment' factor
data_melted$Treatment <- factor(data_melted$Treatment, levels = c("Control", "Sterile feed", "Antibiotics (2X)", "Antibiotics (1X)"))

# Summarize data
Epi_data_summarized <- data_melted %>%
  group_by(Treatment, variable) %>%
  summarise(sum_value = mean(value))


# Function to compute Kruskal-Wallis test and extract p-value
compute_kruskal_pvalue <- function(data) {
  kruskal_result <- kruskal.test(sum_value ~ Treatment, data = data)
  p_value <- broom::tidy(kruskal_result)$p.value
  return(p_value)
}

# Compute Kruskal-Wallis p-values for each Treatment and variable combination
p_values_list <- Epi_data_summarized %>%
  group_by(variable) %>%
  do(p_value = compute_kruskal_pvalue(.)) %>%
  ungroup()

# Expand the list of p_values
p_values_list_expanded <- p_values_list %>%
  mutate(p_value = unlist(p_value))

# Print the result
print(p_values_list_expanded)



#### Statistical test
# List of variables
variable_names <- c("Observed", "Chao1", "Shannon", "Simpson", "InvSimpson", "ACE")

# Initialize an empty data frame to store results
data_frames <- data.frame()

# Loop through each variable
for (var_name in variable_names) {
  PT <- Epi_data_summarized %>%
    filter(variable == var_name) %>%
    dunnTest(sum_value ~ Treatment, data = ., method = "bh")
  PT2 <- PT$res
  
  letter_grouping <- cldList(comparison = PT2$Comparison, p.value = PT2$P.adj, threshold = 0.05)
  
  # Create a data frame with the desired column
  new_data <- data.frame(variable = rep(var_name, nrow(letter_grouping)))
  
  # Combine the new data frame with the existing output data frame
  result_df <- cbind(new_data, letter_grouping)
  
  data_frames <- bind_rows(data_frames, result_df)
}

data_frames
# Now data_frames contains the combined results for all variables

# Rename Group to Treatment
colnames(data_frames)[2] ="Treatment"


# Define the renaming changes
change <- c("Antibiotics(1X)" = "Antibiotics (1X)", 
            "Antibiotics(2X)" = "Antibiotics (2X)", 
            "Sterilefeed" = "Sterile feed")

# Apply the changes to the "Treatment" column
data_frames <- data_frames %>%
  mutate(Treatment = ifelse(Treatment %in% names(change), change[Treatment], Treatment))

# Print the updated data frame
print(data_frames)

data_frames$Treatment <- factor(data_frames$Treatment, levels = c("Control", "Sterile feed", "Antibiotics (2X)", "Antibiotics (1X)"))

data_frames$variable <- factor(data_frames$variable, levels = c("Observed", "Chao1", "Shannon", "Simpson", "InvSimpson", "ACE"))

# Print the updated data frame
print(data_frames)

data_frames %>%
  mutate(Treatment = factor(Treatment, levels = c("Control", "Sterile feed", "Antibiotics (2X)", "Antibiotics (1X)"))) %>%
  kable(digits = c(0, 1, 1, 1, 0, 1)) %>%
  kable_styling(bootstrap_options = c("striped", "hover", "condensed", "responsive"), full_width = FALSE)

p_Epi_hindgut +
 geom_label(data = data_frames, aes(x = Treatment, y = -Inf, label = Letter, group = Treatment), 
             vjust = 0.1, inherit.aes = FALSE)

ggsave("Epi_hindgut_alpha_diversity.svg",width = 20, height = 35, units = "cm")
```
